Remove client
